Define fixture in conftest
